requirements.txt
pymongo[srv]
fastapi
apscheduler
cohere
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
Jinja2
uvicorn